+++ DolBlazor/Shared/NavMenu.razor
@@ -25,8 +25,8 @@
         @if (!isAuthenticated)
         {
             <li class="nav-item px-3">
-                <NavLink class="nav-link" href="travel">
-                    <span class="oi oi-compass" aria-hidden="true"></span> Travel
+                <NavLink class="nav-link" href="game">
+                    <span class="oi oi-infinity" aria-hidden="true"></span> Game
                 </NavLink>
             </li>
             <li class="nav-item px-3">
